Correct behaviour of colour overlay?
I believe this is what one would expect to happen when padding is disabled? Since the overlay looks a bit weird on square corners and such.
This basically makes the height of the overlay which fades to the top to be exactly equal to the height (and position) of the bar itself.

Forgive me if I'm missing a configuration part or something like that which actually fixes this, but I find it weird as default behaviour.